Performed multiple data adjustments and attempted work with LogisticRegression poor results
Now the model results will be saved in a separated file for archiving and retreating purposes.

All the best_params_ of GridSearchCV will be saved in a txt file, if the process will continue might change that to csv.

Worked on fixing the skew() of train and test datasets via np.log1p, however, the results are still unsatisfactory, FN is still greater than TN.